Fixed warnings

+++ 2D Engine/2D Engine/ColliderComponent.h
@@ -48,8 +48,8 @@ public:
 			collider.x = static_cast<int>(transform->position.x);
 			collider.y = static_cast<int>(transform->position.y);
 
-			collider.w = transform->width * transform->scale;
-			collider.h = transform->height * transform->scale;
+			collider.w = static_cast<int>(transform->width * transform->scale);
+			collider.h = static_cast<int>(transform->height * transform->scale);
 		}
 		destinationRect.x = collider.x - Game::camera.x;
 		destinationRect.y = collider.y - Game::camera.y;
