+++ 2D Engine/2D Engine/PongScene.cpp
@@ -34,24 +34,24 @@ void PongScene::OnEnter(SceneContext& ctx)
 	//Setup entity texture ids in pars with the controls
 	std::map<SDL_Scancode, const char*> leftPaddleKeyboardControlsMap = { {SDL_SCANCODE_W, "paddle"}, {SDL_SCANCODE_S, "paddle"} };
 
-	leftPaddle->AddComponent<TransformComponent>(50, 320, 40, 5, 2);
+	leftPaddle->AddComponent<TransformComponent>(50.0f, 320.0f, 40, 5, 2.0f);
 	leftPaddle->AddComponent<SpriteComponent>(assets, "paddle", false);
 	leftPaddle->AddComponent<ColliderComponent>("Player 1");
 	leftPaddle->AddComponent<KeyboardController>(assets, Game::keyState, leftPaddleKeyboardControlsMap, false, true, InputScheme::WASD);
 	leftPaddle->AddGroup(GroupLabels::groupPlayers);
 
-	rightPaddle->AddComponent<TransformComponent>(750, 320, 40, 5, 2);
+	rightPaddle->AddComponent<TransformComponent>(750.0f, 320.0f, 40, 5, 2.0f);
 	rightPaddle->AddComponent<SpriteComponent>(assets, "paddle", false);
 	rightPaddle->AddComponent<ColliderComponent>("Player 2");
 	rightPaddle->AddComponent<KeyboardController>(assets, Game::keyState, leftPaddleKeyboardControlsMap, false, true, InputScheme::Arrows);
 	rightPaddle->AddGroup(GroupLabels::groupPlayers);
 
-	ball->AddComponent<TransformComponent>(450, 320, 5, 5, 2);
+	ball->AddComponent<TransformComponent>(450.0f, 320.0f, 5, 5, 2.0f);
 	ball->AddComponent<SpriteComponent>(assets, "paddle", false);
 	ball->AddComponent<ColliderComponent>("ball");
 	ball->AddGroup(GroupLabels::groupProjectiles);
 
-	ball->GetComponent<TransformComponent>().velocity = Vector2D(1, 0.5f);
+	ball->GetComponent<TransformComponent>().velocity = Vector2D(1.0f, 0.5f);
 
 
 	SDL_Color whiteColor = { 255,255,255,255 };
@@ -159,14 +159,14 @@ void PongScene::Update(SceneContext& ctx)
 			Vector2D normalVector = Vector2D(-1, 0);
 			Vector2D reflection = ballVelocity.Reflection(normalVector);
 			float newSpeed = CalculateNewSpeed(1.80f);
-			ball->GetComponent<TransformComponent>().velocity = reflection * newSpeed;
+			ball->GetComponent<TransformComponent>().velocity = reflection * static_cast<int>(newSpeed);
 			std::cout << "Ball velocity after collision: " << ball->GetComponent<TransformComponent>().velocity << std::endl;
 		}
 		if (Collision::AABB(leftPaddle->GetComponent<ColliderComponent>().collider, ball->GetComponent<ColliderComponent>().collider)) {
 			Vector2D normalVector = Vector2D(-1, 0);
 			Vector2D reflection = ballVelocity.Reflection(normalVector);
 			float newSpeed = CalculateNewSpeed(1.80f);
-			ball->GetComponent<TransformComponent>().velocity = reflection * newSpeed;
+			ball->GetComponent<TransformComponent>().velocity = reflection * static_cast<int>(newSpeed);
 			std::cout << "Ball velocity after collision: " << ball->GetComponent<TransformComponent>().velocity << std::endl;
 		}
 
@@ -193,19 +193,19 @@ void PongScene::Update(SceneContext& ctx)
 
 		if (collidesLeft) {
 			rightScore++;
-			leftPaddle->GetComponent<TransformComponent>().position = Vector2D(50, 320);
-			rightPaddle->GetComponent<TransformComponent>().position = Vector2D(750, 320);
-			ball->GetComponent<TransformComponent>().position = Vector2D(450, 320);
+			leftPaddle->GetComponent<TransformComponent>().position = Vector2D(50.0f, 320.0f);
+			rightPaddle->GetComponent<TransformComponent>().position = Vector2D(750.0f, 320.0f);
+			ball->GetComponent<TransformComponent>().position = Vector2D(450.0f, 320.0f);
 			currentSpeed = baseSpeed;
-			ball->GetComponent<TransformComponent>().velocity = Vector2D(1, -0.5f).Normalize() * currentSpeed;
+			ball->GetComponent<TransformComponent>().velocity = Vector2D(1.0f, -0.5f).Normalize() * static_cast<int>(currentSpeed);
 		}
 		else if (collidesRight) {
 			leftScore++;
-			leftPaddle->GetComponent<TransformComponent>().position = Vector2D(50, 320);
-			rightPaddle->GetComponent<TransformComponent>().position = Vector2D(750, 320);
-			ball->GetComponent<TransformComponent>().position = Vector2D(450, 320);
+			leftPaddle->GetComponent<TransformComponent>().position = Vector2D(50.0f, 320.0f);
+			rightPaddle->GetComponent<TransformComponent>().position = Vector2D(750.0f, 320.0f);
+			ball->GetComponent<TransformComponent>().position = Vector2D(450.0f, 320.0f);
 			currentSpeed = baseSpeed;
-			ball->GetComponent<TransformComponent>().velocity = Vector2D(1, -0.5f).Normalize() * currentSpeed;
+			ball->GetComponent<TransformComponent>().velocity = Vector2D(1.0f, -0.5f).Normalize() * static_cast<int>(currentSpeed);
 		}
 
 		if (leftScore >= maxScore) {
@@ -247,11 +247,11 @@ void PongScene::ResetGame()
 	rightScore = 0;
 	leftScore = 0;
 
-	leftPaddle->GetComponent<TransformComponent>().position = Vector2D(50, 320);
-	rightPaddle->GetComponent<TransformComponent>().position = Vector2D(750, 320);
-	ball->GetComponent<TransformComponent>().position = Vector2D(450, 320);
+	leftPaddle->GetComponent<TransformComponent>().position = Vector2D(50.0f, 320.0f);
+	rightPaddle->GetComponent<TransformComponent>().position = Vector2D(750.0f, 320.0f);
+	ball->GetComponent<TransformComponent>().position = Vector2D(450.0f, 320.0f);
 
-	ball->GetComponent<TransformComponent>().velocity = Vector2D(1, -0.5);
+	ball->GetComponent<TransformComponent>().velocity = Vector2D(1.0f, -0.5f);
 
 	if (gameOver) {
 		gameOver->Close();
